Update concept-best-practices-scanning.md
Description of assigning assets to subcollections using scans at line 32 is confusing.  Scans are created for the data source and not team subcollections.  

I also do not understand this sentence: "Sources without parents are grouped in a dotted box in the map view with no arrows linking them to parents."  The diagram clearly shows a line from a parent collection to the dotted box containing a registered data source.